+++ src/Cacheresource/File.php
@@ -19,10 +19,7 @@ use Smarty\Template\Cached;
 
 /**
  * This class does contain all necessary methods for the HTML cache on file system
- * Implements the file system as resource for the HTML cache Version ussing nocache inserts.
- *
-
-
+ * Implements the file system as resource for the HTML cache Version using nocache inserts.
  */
 class File extends Base
 {