+++ src/Compile/Tag/NocacheClose.php
@@ -35,7 +35,6 @@ class NocacheClose extends Base {
 		[$compiler->nocache] = $this->closeTag($compiler, ['nocache']);
 		// this tag does not return compiled code
 		$compiler->has_code = false;
-		echo "has_code set to false in " . __METHOD__ . "\n";
 		return true;
 	}
 }