+++ src/FunctionHandler/FunctionHandlerInterface.php
@@ -7,5 +7,4 @@ use Smarty\Template;
 interface FunctionHandlerInterface {
 	public function handle($params, Template $template);
 	public function isCacheable(): bool;
-	public function getCacheAttributes(): array;
 }